moved polarity class into aminoacid
