+++ deeprankcore/models/amino_acid.py
@@ -1,6 +1,22 @@
 import numpy
+from enum import Enum
 
-from deeprankcore.models.polarity import Polarity
+
+
+class Polarity(Enum):
+    "a value to express a residue's polarity"
+
+    APOLAR = 0
+    POLAR = 1
+    NEGATIVE_CHARGE = 2
+    POSITIVE_CHARGE = 3
+
+    @property
+    def onehot(self):
+        t = numpy.zeros(4)
+        t[self.value] = 1.0
+
+        return t
 
 
 class AminoAcid:
@@ -100,9 +116,6 @@ class AminoAcid:
     def __repr__(self):
         return self._name
 
-
-from deeprankcore.models.amino_acid import AminoAcid
-from deeprankcore.models.polarity import Polarity
 
 alanine = AminoAcid(
     "Alanine",
